middleware/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
pika==1.3.1
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.0
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
